Move off the device and mode config to a separate file.
