+++ theano/sandbox/gpuarray/tests/test_neighbours.py
@@ -1,7 +1,6 @@
-
 from theano.tensor.nnet.tests import test_neighbours
-# We let that import do the init of the back-end if needed.
-from .test_basic_ops import mode_with_gpu
+
+from .config import mode_with_gpu
 
 from ..neighbours import GpuImages2Neibs
