+++ theano/sandbox/gpuarray/tests/test_opt.py
@@ -14,7 +14,9 @@ from ..basic_ops import (GpuAlloc, GpuReshape, gpu_alloc,
                          gpu_from_host, host_from_gpu)
 from ..elemwise import GpuCAReduceCuda, GpuCAReduceCPY, GpuElemwise
 from ..subtensor import GpuSubtensor
-from .test_basic_ops import rand_gpuarray, mode_with_gpu, mode_without_gpu
+
+from .config import mode_with_gpu, mode_without_gpu
+from .test_basic_ops import rand_gpuarray
 
 
 def test_local_assert():
